Implemented remove course
