+++ app.js
@@ -10,7 +10,6 @@ var upload = multer({ dest: 'public/images' });
 
 var signin = require('./routes/signin');
 var index = require('./routes/index');
-var add = require('./routes/addCourse');
 var myProfile = require('./routes/my-profile');
 var myGroups = require('./routes/my-groups');
 var course = require('./routes/course');
@@ -34,10 +33,12 @@ app.get('/', signin.view);
 app.post('/login', signin.login);
 app.post('/register', signin.registerUser);
 app.get('/index', index.view);
-app.get('/addCourse', add.addCourse);
+app.post('/index/add-course', index.addCourse);
+app.post('/index/remove-course', index.removeCourse);
 app.get('/my-profile', myProfile.view);
 app.post('/my-profile', upload.single('avatar'), myProfile.updateProfile);
 app.get('/my-groups', myGroups.view);
+app.post('/my-groups', myGroups.leaveGroup);
 app.get('/course/:courseId', course.view);
 app.get('/course/:courseId/peer/:peerId', peer.view);
 app.post('/course/:courseId/peer/:peerId', peer.joinGroup);
