+++ routes/index.js
@@ -17,3 +17,31 @@ exports.view = function(req, res) {
     courses : courses
    });
 };
+
+exports.addCourse = function(req, res) { 
+	var user = data.students[0];
+
+	var newCourse = {
+      "id" : "CSE 190",
+      "title" : "Entrepreneurship",
+      "groupSize" : 2,
+      "students" : [user.id],
+	     "modalID" : "4"
+    };
+
+	// Professor/TA action
+	data.courses.push(newCourse);
+	// Student action
+	user.groups[newCourse.id] = [];
+	user.invites[newCourse.id] = [];
+
+	res.redirect('/index');
+ };
+
+exports.removeCourse = function(req, res) {
+  var courseId = req.body.id;
+  var user = data.students[0];
+  delete user.groups[courseId];
+  delete user.invites[courseId];
+  res.redirect('/index');
+};
