+++ views/index.hbs
@@ -37,7 +37,7 @@
 
 	<div class="home">
 	    <h1 class="page-header">Hello, {{name}}</h1>
-	    
+
 	    <div class="row placeholders">
 	    	<h4><center>Pick a course from below to view peers.</center></h4>
 	    </div>
@@ -81,8 +81,8 @@
 	    <div class="row placeholders">
 	      <h4>Add a course</h4>
 	      <p>Enter your course code provided by your professor.<p>
-	      <form id="addCourseForm" role="form" method="get" action="/addCourse">
-	      	<input type="text" placeholder="12345" name="addCourse">
+	      <form id="addCourseForm" role="form" method="post" action="/index/add-course">
+	      	<input type="text" placeholder="12345" name="courseId">
 	      	<input type="submit" value="Submit">
 	      </form>
 	    </div>
